# mobile/openapi/identity-service.yaml
...
openapi: 3.1.0
info:
  title: TD IT Solutions — Identity Service API
  version: "1.0.0"
  description: >
    Feature 001 (Customer Account Creation & Authentication). Mediates all
    client access to Supabase Auth per ADR-0002's mediation principle —
    no client ever holds a Supabase-honoured credential (see api-design.md §1).
servers:
  - url: /v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend-minted access token (api-design.md §1). Never a Supabase-issued JWT.
    internalServiceAuth:
      type: apiKey
      in: header
      name: X-Internal-Service-Key
      description: Service-to-service auth for the internal-only account-status surface (§2.3 row 4). Not reachable by any client.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Client-generated UUID v4. See api-design.md §4.
...
  headers:
...
    RetryAfter:
      schema: { type: integer, description: Seconds. }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            requestId: { type: string, format: uuid }
...
    UserType:
      type: string
      enum: [customer, admin, security_company_operator, support_agent]
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, expired, revoked]
    InvitationPublic:
      type: object
      description: Invitee-facing view — never exposes token_hash.
      properties:
        email: { type: string, format: email }
        userType: { $ref: '#/components/schemas/UserType' }
        partnerOrganizationName: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/InvitationStatus' }
        expiresAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Validation error.
      headers:
        X-Request-Id: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired/revoked bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Idempotency-key reuse with a different body, or a state conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UpstreamUnavailable:
      description: Supabase Admin API unreachable/timed out (api-design.md §6).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /invitations:
    post:
      operationId: createInvitation
      summary: Admin-only. Issues an invitation for admin/security_company_operator/support_agent.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, userType]
              properties:
                email: { type: string, format: email }
                userType:
                  type: string
                  enum: [admin, security_company_operator, support_agent]
                partnerOrganizationId:
                  type: string
                  format: uuid
                  description: Required if userType = security_company_operator (BR-7).
      responses:
        '201':
          description: Invitation created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  status: { $ref: '#/components/schemas/InvitationStatus' }
                  expiresAt: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Caller is not admin (ruling C4 — support agents may not issue invitations).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /invitations/{token}:
    get:
      operationId: getInvitation
      summary: Invitee views invite details before accepting. Token is opaque; backend hashes and looks up server-side (never a direct table query — database-design.md §5.4).
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationPublic' }
        '404':
          description: Token not found, expired, or revoked — same shape for all three (no enumeration of *why* it's invalid).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: >
            Account created, invitation marked accepted. Does NOT issue a
            session — the caller must proceed to mandatory MFA enrollment
            (BR-4) before any session is minted; see /mfa/enroll.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string, format: uuid }
                  mfaEnrollmentRequired: { type: boolean, enum: [true] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: Invitation expired or already accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }

  # ---------------------------------------------------------------
  # MFA enrollment
  # ---------------------------------------------------------------
...